Updated plugin to softdepend CBan, added CBan dependencies,refactored BanManager as abstract and shifted existing implementation to FBan, with new CBan to integrate with CBanManagement.
