+++ pom.xml
@@ -6,7 +6,7 @@
 	<groupId>com.untamedears</groupId>
 	<artifactId>PrisonPearl</artifactId>
 	<packaging>jar</packaging>
-	<version>2.7.1</version>
+	<version>2.7.2</version>
 	<name>PrisonPearl</name>
 	<url>https://github.com/Civcraft/PrisonPearl</url>
 
@@ -71,6 +71,18 @@
 			<groupId>vg.civcraft.mc.bettershards</groupId>
 			<artifactId>BetterShards</artifactId>
 			<version>[1.3.17,)</version>
+			<scope>provided</scope>
+		</dependency>
+		<dependency>
+			<groupId>vg.civcraft.mc</groupId>
+			<artifactId>CBanManagement</artifactId>
+			<version>[1.3.1,)</version>
+			<scope>provided</scope>
+		</dependency>
+		<dependency>
+			<groupId>vg.civcraft.mc.civmodcore</groupId>
+			<artifactId>CivModCore</artifactId>
+			<version>[1.0.1,)</version>
 			<scope>provided</scope>
 		</dependency>
 	</dependencies>
